+++ common/src/main/java/net/mehvahdjukaar/polytone/PlatStuff.java
@@ -218,9 +218,5 @@ public class PlatStuff {
         throw new AssertionError();
     }
 
-    @ExpectPlatform
-    public static RegistryAccess getServerRegistryAccess() {
-        throw new AssertionError();
-    }
 
 }